+++ src/components/breadcrumbs/index.jsx
@@ -1,28 +1,29 @@
-import React from "react";
+import React, { PropTypes } from "react";
 import radium from "radium";
 import capitalize from "lodash/capitalize";
-import { color } from "../../../settings.json";
+import colors from "../../styles/colors";
+import { fontSizeBodySmall, fontWeightBook } from "../../styles/typography";
 import Icon from "../icon";
 import { blueLink } from "../../utils/mixins";
 import schema from "../../utils/schema";
 import font from "../../utils/font";
+import { rgba } from "../../utils/color";
+import propTypes from "../../utils/propTypes";
 
 const _ = { capitalize };
 
 const styles = {
   container: {
-    base: {
-      color: color.lightText,
-      fontFamily: font("benton"),
-      fontSize: "14px",
-      fontWeight: 400,
-      lineHeight: 1,
-    },
+    color: rgba(colors.textPrimary, 0.56),
+    fontFamily: font("benton"),
+    fontSize: `${fontSizeBodySmall}px`,
+    fontWeight: fontWeightBook,
+    lineHeight: 1,
   },
 
   item: {
     padded: {
-      marginRight: `${10 / 14}em`,
+      marginRight: `${10 / fontSizeBodySmall}em`,
     },
   },
 
@@ -30,7 +31,7 @@ const styles = {
     base: blueLink(),
 
     padded: {
-      marginRight: `${10 / 14}em`,
+      marginRight: `${10 / fontSizeBodySmall}em`,
     },
   },
 };
@@ -58,15 +59,12 @@ function linkMicroData(link) {
   });
 }
 
-/**
- * Breadcrumb navigation component
- */
-function Breadcrumbs({ links }) {
+function Breadcrumbs({ links, style }) {
   const items = links.map((link, index) => (
     <span
       className="Breadcrumbs-item"
       style={[index !== links.length - 1 && styles.item.padded]}
-      key={index}
+      key={link.href}
       {...itemMicroData}
     >
       <a
@@ -75,7 +73,7 @@ function Breadcrumbs({ links }) {
           index < links.length - 1 && styles.anchor.padded,
         ]}
         href={`//www.lonelyplanet.com${link.href}`}
-        key={index}
+        key={link.href}
         {...linkMicroData(link)}
       >
         <span itemProp="name">{link.title}</span>
@@ -83,8 +81,8 @@ function Breadcrumbs({ links }) {
 
       {index < links.length - 1 &&
         <Icon.ChevronRight
-          width={`${6 / 14}em`}
-          height={`${6 / 14}em`}
+          width={`${6 / fontSizeBodySmall}em`}
+          height={`${6 / fontSizeBodySmall}em`}
         />
       }
 
@@ -95,7 +93,7 @@ function Breadcrumbs({ links }) {
   return (
     <nav
       className="Breadcrumbs"
-      style={styles.container.base}
+      style={[styles.container, style]}
       {...listMicroData}
     >
       {items}
@@ -104,22 +102,14 @@ function Breadcrumbs({ links }) {
 }
 
 Breadcrumbs.propTypes = {
-  /**
-   * An array of links for the navigation
-   */
-  links: React.PropTypes.arrayOf(
-    React.PropTypes.shape({
-      title: React.PropTypes.string,
-      href: React.PropTypes.string,
-      type: React.PropTypes.string,
+  links: PropTypes.arrayOf(
+    PropTypes.shape({
+      title: PropTypes.string,
+      href: PropTypes.string,
+      type: PropTypes.string,
     }),
   ).isRequired,
+  style: propTypes.style,
 };
-
-Breadcrumbs.defaultProps = {
-  links: [],
-};
-
-Breadcrumbs.styles = styles;
 
 export default radium(Breadcrumbs);